+++ src/main/java/interfacesDAO/AlumnoDAO.java
@@ -8,5 +8,5 @@ import clases.Comentario;
 
 public interface AlumnoDAO extends GenericDAO<Alumno>{
 	List<Comentario> obtenerComentarios(long id);
-	List<Cartelera> obtenerIntereses(int id); // Ver como se guardan los intereses.
+	List<Cartelera> obtenerIntereses(long id); // Ver como se guardan los intereses.
 }
